Move components and containers to their directories
