+++ src/containers/ProfilePictureContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import ProfilePicture from 'src/profile/ProfilePicture';
+import ProfilePicture from 'src/views/ProfilePicture';
 import { pushCameraRollRoute } from 'src/actions/navigation';
 
 export default connect(
