+++ src/views/Activation.js
