+++ src/views/ActivityList/index.js
