+++ src/views/ActivityList/styles.js
