+++ src/views/CameraRoll/__tests__/index-test.js
