+++ src/views/CameraRoll/styles.js
