+++ src/views/Category/styles.js
