+++ src/views/CategoryAbout/__tests__/CategoryAbout-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import CategoryAbout from 'src/activities/CategoryAbout';
+import CategoryAbout from 'src/views/CategoryAbout';
 import { fakeCategory } from 'app/scripts/helpers';
 import images from 'src/constants/images';
