+++ src/views/NavigationStack.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react';
 import { NavigationExperimental } from 'react-native';
 const { CardStack: NavigationCardStack } = NavigationExperimental;
 import NotFound from 'src/components/NotFound';
-import ROUTES_COMPONENTS from 'src/navigation/routes/actions/components';
+import ROUTES_COMPONENTS from 'src/routes';
 
 
 const router = key => {
