+++ src/views/ProfilePicture/__tests__/index.js
