+++ src/views/ScheduleDetail.js
